refactor attributes directory structure
