+++ packages/core/src/attributes/common/attribute.ts
