+++ packages/core/src/attributes/common/contracts.ts
