+++ packages/core/src/attributes/types/classes/Attr.ts
