+++ packages/core/src/attributes/types/classes/Number.ts
