+++ packages/core/src/attributes/types/classes/String.ts
