+++ packages/core/src/attributes/types/createAttrField.ts
@@ -1,5 +1,5 @@
-import type { ModelConstructor } from '../model/types'
-import { Attr } from './classes/types/Attr'
+import type { ModelConstructor } from '../../model/types'
+import { Attr } from './classes/Attr'
 
 export const createAttrField = (model: ModelConstructor<any>, value?: any) => {
   return new Attr(model.newRawInstance(), value)
