+++ packages/core/src/attributes/types/createBooleanField.ts
@@ -1,5 +1,5 @@
-import type { ModelConstructor } from '../model/types'
-import { Boolean } from './classes/types/Boolean'
+import type { ModelConstructor } from '../../model/types'
+import { Boolean } from './classes/Boolean'
 
 export const createBooleanField = (model: ModelConstructor<any>, value: boolean | null) => {
   return new Boolean(model.newRawInstance(), value)
