+++ packages/core/src/attributes/types/createNumberField.ts
@@ -1,5 +1,5 @@
-import type { ModelConstructor } from '../model/types'
-import { Number } from './classes/types/Number'
+import type { ModelConstructor } from '../../model/types'
+import { Number } from './classes/Number'
 
 export const createNumberField = (model: ModelConstructor<any>, value: number | null) => {
   return new Number(model.newRawInstance(), value)
