+++ packages/core/src/attributes/types/createStringField.ts
@@ -1,5 +1,5 @@
-import type { ModelConstructor } from '../model/types'
-import { String } from './classes/types/String'
+import type { ModelConstructor } from '../../model/types'
+import { String } from './classes/String'
 
 export const createStringField = (model: ModelConstructor<any>, value: string | null) => {
   return new String(model.newRawInstance(), value)
