+++ packages/core/src/attributes/types/createUidField.ts
@@ -1,5 +1,5 @@
-import type { ModelConstructor } from '../model/types'
-import { Uid } from './classes/types/Uid'
+import type { ModelConstructor } from '../../model/types'
+import { Uid } from './classes/Uid'
 
 export const createUidField = (model: ModelConstructor<any>, value?: any) => {
   return new Uid(model.newRawInstance(), value)
