+++ packages/core/src/decorators.ts
@@ -1,6 +1,6 @@
-export { BelongsTo } from '@/attributes/decorators/attributes/relations/BelongsTo'
-export { HasMany } from '@/attributes/decorators/attributes/relations/HasMany'
-export { HasManyBy } from '@/attributes/decorators/attributes/relations/HasManyBy'
-export { HasOne } from '@/attributes/decorators/attributes/relations/HasOne'
-export { MorphOne } from '@/attributes/decorators/attributes/relations/MorphOne'
-export { MorphTo } from '@/attributes/decorators/attributes/relations/MorphTo'
+export { BelongsTo } from '@/attributes/relations/decorators/BelongsTo'
+export { HasMany } from '@/attributes/relations/decorators/HasMany'
+export { HasManyBy } from '@/attributes/relations/decorators/HasManyBy'
+export { HasOne } from '@/attributes/relations/decorators/HasOne'
+export { MorphOne } from '@/attributes/relations/decorators/MorphOne'
+export { MorphTo } from '@/attributes/relations/decorators/MorphTo'
