+++ packages/core/src/index.ts
@@ -12,11 +12,11 @@ export * from './query/types'
 export * from './repository/repository'
 export * from './schema/schema'
 export * from './types/index'
-export { BelongsTo as BelongsToAttr } from '@/attributes/classes/relations/belongs-to'
-export { HasMany as HasManyAttr } from '@/attributes/classes/relations/has-many'
-export { HasManyBy as HasManyByAttr } from '@/attributes/classes/relations/has-many-by'
-export { HasOne as HasOneAttr } from '@/attributes/classes/relations/has-one'
-export { MorphOne as MorphOneAttr } from '@/attributes/classes/relations/morph-one'
-export { MorphTo as MorphToAttr } from '@/attributes/classes/relations/morph-to'
-export * from '@/attributes/classes/relations/relation'
+export { BelongsTo as BelongsToAttr } from '@/attributes/relations/classes/belongs-to'
+export { HasMany as HasManyAttr } from '@/attributes/relations/classes/has-many'
+export { HasManyBy as HasManyByAttr } from '@/attributes/relations/classes/has-many-by'
+export { HasOne as HasOneAttr } from '@/attributes/relations/classes/has-one'
+export { MorphOne as MorphOneAttr } from '@/attributes/relations/classes/morph-one'
+export { MorphTo as MorphToAttr } from '@/attributes/relations/classes/morph-to'
+export * from '@/attributes/relations/classes/relation'
 export { Schemas } from '@/schema/types'
