+++ packages/core/src/model/Model.ts
@@ -1,13 +1,13 @@
 import { isUnknownRecord } from '@core-shared-utils/isUnknownRecord'
 
-import { BelongsTo } from '@/attributes/classes/relations/belongs-to'
-import { HasMany } from '@/attributes/classes/relations/has-many'
-import { HasManyBy } from '@/attributes/classes/relations/has-many-by'
-import { HasOne } from '@/attributes/classes/relations/has-one'
-import { MorphOne } from '@/attributes/classes/relations/morph-one'
-import { MorphTo } from '@/attributes/classes/relations/morph-to'
-import { Relation } from '@/attributes/classes/relations/relation'
 import type { Attribute, Type } from '@/attributes/field-types'
+import { BelongsTo } from '@/attributes/relations/classes/belongs-to'
+import { HasMany } from '@/attributes/relations/classes/has-many'
+import { HasManyBy } from '@/attributes/relations/classes/has-many-by'
+import { HasOne } from '@/attributes/relations/classes/has-one'
+import { MorphOne } from '@/attributes/relations/classes/morph-one'
+import { MorphTo } from '@/attributes/relations/classes/morph-to'
+import { Relation } from '@/attributes/relations/classes/relation'
 import type { Collection, Element, Item, RawModel } from '@/data/types'
 import type { ModelConstructor } from '@/model/types'
 import { assert, isArray, isFunction, isNullish } from '@/support/utils'
