+++ packages/core/test/model/Model_Attrs_Number.spec.ts
@@ -1,4 +1,4 @@
-import { NumberField } from '../../src/attributes/decorators/attributes/types/NumberField'
+import { NumberField } from '../../src/attributes/types/decorators/NumberField'
 import { Model } from '@/model/Model'
 
 describe('unit/model/Model_Attrs_Number', () => {
