+++ packages/core/test/model/Model_Attrs_UID.spec.ts
@@ -1,6 +1,6 @@
 import { mockUid } from '../../../../test/utils/Helpers'
 
-import { UidField } from '../../src/attributes/decorators/attributes/types/UidField'
+import { UidField } from '../../src/attributes/types/decorators/UidField'
 import { Model } from '@/model/Model'
 
 describe('unit/model/Model_Attrs_UID', () => {
